+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 0.7.2 (unreleased)
 * Handle non-existent files/directories gracefully
 * Teach scss-lint that opening curly braces should be preceded by one space
+* Teach scss-lint that placeholder names should be lowercase and use hyphens
 
 0.7.1
 * Upgrade Sass gem dependency from 3.2.8 -> 3.2.9